pyproject.toml
[tool.poetry]
name = "speech"
version = "0.1.0"
description = ""
authors = ["Aurimas Pabrinkis <hello@apbr.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
torch = "^2.2.0"
torchaudio = "^2.2.0"
uvicorn = "^0.27.0.post1"
sentencepiece = "^0.1.99"
transformers = {git = "https://github.com/huggingface/transformers"}
scipy = "^1.12.0"
python-multipart = "^0.0.9"
pytube = "^15.0.0"
openai = "^1.12.0"
pydantic = "^2.6.1"
webvtt-py = "^0.4.6"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
